requirements.txt
aerich==0.5.8
aiogram==3.21.0
aiohttp
loguru==0.6.0
openpyxl==3.1.5
pandas==2.3.1
polib==1.2.0
python-dotenv==0.19.2
pytz==2024.1
redis==4.5.4
requests==2.32.3
tortoise-orm==0.17.8
asyncpg
structlog
Babel
